requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.2.2+cpu
torchvision==0.17.2+cpu
torchaudio==2.2.2+cpu

fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.9

ultralytics==8.4.12
pillow==11.1.0
onnxruntime==1.23.2
pydantic==2.10.6

numpy==1.26.4